Closes #223
Had a lot of trouble trying to figure out how to return the player as ARunner* instead of AActor*. As far as I can tell, it's not possible with the current implementation as Runner.h requires AISpawnerController and if AISpawnerController requires Runner, there's a circular declaration that doesn't work in C++ (as far as I know). So, that being said this still works but when calling get player here you will need to cast it to ARunner.